feat: dont delete head branch unless pr is closed or merged

+++ plugins/aladino/actions/deleteHeadBranch.go
@@ -5,6 +5,8 @@
 package plugins_aladino_actions
 
 import (
+	"errors"
+
 	"github.com/reviewpad/reviewpad/v3/codehost/github/target"
 	"github.com/reviewpad/reviewpad/v3/handler"
 	"github.com/reviewpad/reviewpad/v3/lang/aladino"
@@ -21,6 +23,10 @@ func DeleteHeadBranch() *aladino.BuiltInAction {
 func deleteHeadBranch(e aladino.Env, args []aladino.Value) error {
 	target := e.GetTarget().(*target.PullRequestTarget)
 	targetEntity := target.GetTargetEntity()
+
+	if !*target.PullRequest.Merged && target.PullRequest.ClosedAt == nil {
+		return errors.New("pull request should be merged or closed before deleting head branch")
+	}
 
 	return e.GetGithubClient().DeleteReference(e.GetCtx(), targetEntity.Owner, targetEntity.Repo, "heads/"+*target.PullRequest.Head.Ref)
 }
